--- index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Google Maps Demo</title>
-    <button id="toggleBlmOverlay">Toggle BLM Overlay</button>
-    <script src="https://polyfill.io/v3/polyfill.min.js?features=default"></script>
-    <script defer type="module" src="index.js"></script>
-    <link rel="stylesheet" href="styles.css" />
-  </head>
-  <body>
-    <h3>My Google Maps Demo</h3>
-    <input
-      id="pac-input"
-      class="controls"
-      type="text"
-      placeholder="Search Box"
-    />
-
-    <!-- The div element for the map -->
-    <div id="map"></div>
-
-    <button id="locationButton">Get Current Location</button>
-  </body>
-</html>
