+++ package.json
@@ -9,7 +9,15 @@
   "author": "me",
   "license": "ISC",
   "dependencies": {
+    "@googlemaps/google-maps-services-js": "^3.4.0",
     "@googlemaps/js-api-loader": "^1.16.6",
-    "http-server": "^14.1.1"
+    "axios": "^1.7.2",
+    "core-js": "^3.37.1",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "nodejs": "^0.0.0",
+    "path": "^0.12.7",
+    "request": "^2.88.2"
   }
 }
