+++ public/cellmap.png
