+++ public/markerModal.html
@@ -1,0 +1,35 @@
+<div class="info-window-content">
+  <link rel="stylesheet" href="markerModal.css" />
+
+  <h4>Marker Name <button id="editMarker">edit</button></h4>
+
+  <!--button class="getWeather">Get Weather</!--button>-->
+
+  <div id="requestedWeather"></div>
+
+  <p>
+    Elevation: 10,420 ft.
+    <!-- ELEVATION_PLACEHOLDER -->
+  </p>
+  <p>
+    Lat: 110.2343
+    <!-- LATITUDE_PLACEHOLDER -->
+  </p>
+  <p>
+    Lng: -109.3454
+    <!-- LONGITUDE_PLACEHOLDER -->
+  </p>
+
+  <input autofocus required placeholder="Marker name" hidden /> <br />
+
+  <p class="markerNotes">Your marker information</p>
+
+  <input placeholder="Notes / comments" class="notes" hidden /> <br />
+
+  <div id="fileInput"><input type="file" id="myFile" name"filename"></div>
+
+  <button name="submit">Save</button>
+  <div id="delete">
+    <button name="delete" class="delete">X</button>
+  </div>
+</div>
