--- styles.css
@@ -1,4 +1,0 @@
-#map {
-  height: 75vh;
-  width: 100%;
-}
